+++ src/main/java/com/test/bot/repository/UserRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.test.bot.models.UserEntity;
 
 @Repository
-public interface UserRepository extends JpaRepository<UserEntity, Integer>{
-    Boolean existsBytgId(int tgId);
+public interface UserRepository extends JpaRepository<UserEntity, Long>{
+    Boolean existsBytgId(Long tgId);
 }
